Some big refactoring:
use defer to ensure unlock always happens
use the open file descriptor locking (works on linux only) - the posix
locks use (inode, pid) for locking which is common across threads, so it
does not working for synchronization acorss threads. Since in golang it
is more common to use goroutines instead of multiple processes for
concurrency, it is important we use thread friendly locking. For other
operating systems it is possible to write a wrapper to support this
(maybe in future).
The linux only open file descriptor locking uses the file descriptor for
locking. So as long as each thread has its own fd to access the
database, they can cooperate and work.